refactor image viewer libraries.
