+++ pyviewerlib/core/image_viewer/__init__.py
@@ -1,25 +1,17 @@
 import os
-import sys
 import tempfile
 import subprocess
 import mimetypes
-from pathlib import Path
 from typing import Any, Union, Optional
+from importlib import import_module
 
-try:
-    from screeninfo import get_monitors
-except ImportError:
-    get_screen = False
-else:
-    get_screen = True
-
-from . import args_chk, debug_print, debug, json_opts, Args
-sys.path.append(str(Path(__file__).parent.parent.parent))
+from .. import args_chk, debug_print, debug, json_opts, Args
 from pymeflib.color import make_bitmap
 from pymeflib.util import chk_cmd
 
 # image viewer
-img_viewer = None
+__ImgViewer = None
+__mods = ['PIL', 'matplotlib', 'OpenCV']
 
 
 def get_image_viewer(args: Args) -> Optional[str]:
@@ -37,67 +29,43 @@ def get_image_viewer(args: Args) -> Optional[str]:
     str
         the name of image viewer.
     """
-    global img_viewer
-    if img_viewer is not None:
-        return img_viewer
+    global __ImgViewer
+    if __ImgViewer is not None:
+        return __ImgViewer
     if args_chk(args, 'image_viewer'):
         debug_print('set image viewer from args')
-        img_viewer = args.image_viewer
-        if img_viewer == 'PIL':
-            from PIL import Image
-            global Image
-        elif img_viewer == 'matplotlib':
-            import matplotlib.pyplot as plt
-            global plt
-        elif img_viewer == 'OpenCV':
-            import cv2
-            global cv2
-        else:
-            # external command
-            if not chk_cmd(img_viewer, debug):
-                img_viewer = None
+        __ImgViewer = args.image_viewer
     else:
         debug_print('search available image_viewer')
         try:
             from PIL import Image
             debug_print(' => image_viewer: PIL')
-            global Image
         except ImportError:
             pass
         else:
-            img_viewer = 'PIL'
+            __ImgViewer = 'PIL'
 
-        if img_viewer is None:
+        if __ImgViewer is None:
             try:
                 import matplotlib.pyplot as plt
                 debug_print(' => image_viewer: matplotlib')
-                global plt
             except ImportError:
                 pass
             else:
-                img_viewer = 'matplotlib'
+                __ImgViewer = 'matplotlib'
 
-        if img_viewer is None:
+        if __ImgViewer is None:
             try:
                 import cv2
                 debug_print(' => image_viewer: OpenCV')
-                global cv2
             except ImportError:
-                debug_print("can't find image_viewer")
-                img_viewer = None
+                pass
             else:
-                img_viewer = 'OpenCV'
-    return img_viewer
+                __ImgViewer = 'OpenCV'
 
-
-def clear_mpl_axes(axes):
-    # not display axes
-    axes.xaxis.set_visible(False)
-    axes.yaxis.set_visible(False)
-    axes.spines['top'].set_visible(False)
-    axes.spines['bottom'].set_visible(False)
-    axes.spines['right'].set_visible(False)
-    axes.spines['left'].set_visible(False)
+        if __ImgViewer is None:
+            debug_print("can't find image_viewer")
+    return __ImgViewer
 
 
 def get_exec_cmds(args, fname):
@@ -129,32 +97,21 @@ def show_image_file(img_file: str, args: Args) -> bool:
     bool
         Return True if the file opened successfully, otherwise False.
     """
-    name = os.path.basename(img_file)
     img_viewer = get_image_viewer(args)
     debug_print('  use {}'.format(img_viewer))
     if not os.path.isfile(img_file):
         debug_print('image file {} in not found'.format(img_file))
         return False
     if img_viewer is None:
-        print("I can't find any libraries to show image. Please install Pillow or matplotlib.")
+        print("I can't find any libraries to show image.")
         return False
-    elif img_viewer == 'PIL':
-        with Image.open(img_file) as image:
-            image.show(title=name)
-    elif img_viewer == 'matplotlib':
-        img = plt.imread(img_file)
-        fig1 = plt.figure()
-        ax11 = fig1.add_axes((0, 0, 1, 1))
-        ax11.imshow(img)
-        clear_mpl_axes(ax11)
-        plt.show()
-        plt.close(fig1)
-    elif img_viewer == 'OpenCV':
-        img = cv2.imread(img_file)
-        cv2.imshow(name, img)
-        cv2.waitKey(0)
-        cv2.destroyWindow(name)
+    elif img_viewer in __mods:
+        mod = import_module(f'pyviewerlib.core.image_viewer.{img_viewer}')
+        mod.show_image_file(img_file)
     else:
+        if not chk_cmd(img_viewer):
+            print(f'{img_viewer} is not executable')
+            return False
         cmds = get_exec_cmds(args, img_file)
         subprocess.run(cmds)
         # wait to open file. this is for, e.g., open command on Mac OS.
@@ -186,36 +143,13 @@ def show_image_ndarray(data: Any, name: str, args: Args) -> bool:
     if img_viewer is None:
         print("I can't find any libraries to show image. Please install Pillow or matplotlib.")
         return False
-    elif img_viewer == 'PIL':
-        with Image.fromarray(data) as image:
-            image.show(title=name)
-    elif img_viewer == 'matplotlib':
-        if get_screen:
-            height = get_monitors()[0].height
-        else:
-            height = 540
-        rate = data.shape[0]/height*100
-        h = int(data.shape[0]/rate)
-        w = int(data.shape[1]/rate)
-        fig1 = plt.figure(figsize=(w, h))
-        # full display
-        ax1 = fig1.add_axes((0, 0, 1, 1))
-        ax1.imshow(data)
-        clear_mpl_axes(ax1)
-        plt.show()
-        plt.close(fig1)
-    elif img_viewer == 'OpenCV':
-        if data.shape[2] == 3:
-            img = data[:, :, ::-1]  # RGB -> BGR
-        elif data.shape[2] == 4:
-            img = data[:, :, [2, 1, 0, 3]]  # RGBA -> BGRA
-        else:
-            print('invalid data shape')
-            return False
-        cv2.imshow(name, img)
-        cv2.waitKey(0)
-        cv2.destroyWindow(name)
+    elif img_viewer in __mods:
+        mod = import_module(f'pyviewerlib.core.image_viewer.{img_viewer}')
+        mod.show_image_ndarray(data, name)
     else:
+        if not chk_cmd(img_viewer):
+            print(f'{img_viewer} is not executable')
+            return False
         with tempfile.NamedTemporaryFile(suffix='.bmp') as tmp:
             make_bitmap(tmp.name, data, verbose=debug)
             cmds = get_exec_cmds(args, tmp.name)
